requirements.txt
Django>=5.2,<6.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-decouple
django-filter
gunicorn
whitenoise
dj-database-url
psycopg2-binary
pytest-django